ai-safety-auditing/requirements.txt
# LLM APIs
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Statistical Analysis
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.25.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
requests>=2.31.0

# FastAPI & Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
